Component Spline Base
Created a Spline Map to show off an object moving along a Spline. At the moment, it's very basic. Scene starts and the object move along the spline once and done, but it shows the basic.

Rest needs to work with events, we need to be able to easily visualise the points of the spline so it's easier to see the coordinates.

Big thing will be Events Triggered along the spline. Different ways to go about it, but probably using collision system as it's easier and reliable